requirements.txt
discord.py==2.0.1
boto3==1.24.61
httpx==0.23.0
types-requests==2.28.9
filetype==1.1.0
fuzzywuzzy[speedup]==0.18.0
cachetools==5.2.0
types-cachetools==5.2.0
